+++ app/Core/Database/MigrationStatus.php
@@ -122,6 +122,10 @@ final class MigrationStatus
 		return $this->isMigrationCompleted(Migrations::CREATED_DAY);
 	}
 
+	public function idActionIndexCompleted(): bool {
+		return $this->isMigrationCompleted(Migrations::ID_ACTION_INDEX);
+	}
+
 	public function mailLogDataRunning(): bool {
 		return $this->isMigrationRunning(Migrations::MAIL_LOG_DATA);
 	}
@@ -132,6 +136,10 @@ final class MigrationStatus
 
 	public function createdDayRunning(): bool {
 		return $this->isMigrationRunning(Migrations::CREATED_DAY);
+	}
+
+	public function idActionIndexRunning(): bool {
+		return $this->isMigrationRunning(Migrations::ID_ACTION_INDEX);
 	}
 
 	public function getAllMigrationStates(): array {