+++ app/Inventory/Migrations.php
@@ -14,6 +14,7 @@ use App\Core\Database\AbstractMigration;
 use App\Core\Database\Migrations\MailRecipientsMigration;
 use App\Core\Database\Migrations\CreatedDayMigration;
 use App\Core\Database\Migrations\MailLogDataMigration;
+use App\Core\Database\Migrations\IdActionIndex;
 
 use Illuminate\Database\Capsule\Manager as Capsule;
 
@@ -26,47 +27,55 @@ class Migrations
 	public const string MAIL_RECIPIENTS = '20260111_mail_recipients';
 	public const string CREATED_DAY = '20260729_created_day';
 	public const string MAIL_LOG_DATA = '20260731_mail_log_data';
+	public const string ID_ACTION_INDEX = '20260806_id_action_index';
 
 	public const array MIGRATIONS = [
 		self::MAIL_RECIPIENTS,
 		self::CREATED_DAY,
 		self::MAIL_LOG_DATA,
+		self::ID_ACTION_INDEX,
 	];
 
 	public const array REQUIRED = [
 		self::MAIL_RECIPIENTS,
 		self::CREATED_DAY,
 		self::MAIL_LOG_DATA,
+		self::ID_ACTION_INDEX,
 	];
 
 	public const array MIGRATION_CLASSES = [
 		self::MAIL_RECIPIENTS => MailRecipientsMigration::class,
 		self::CREATED_DAY => CreatedDayMigration::class,
 		self::MAIL_LOG_DATA => MailLogDataMigration::class,
+		self::ID_ACTION_INDEX => IdActionIndex::class,
 	];
 
 	public const array MIGRATION_DESCR = [
 		self::MAIL_RECIPIENTS => "Mail Log Recipients",
 		self::CREATED_DAY => "Mail Log Created Day",
 		self::MAIL_LOG_DATA => "Mail Log Data",
+		self::ID_ACTION_INDEX => "id action Index",
 	];
 
 	public const array MIGRATION_BATCH = [
 		self::MAIL_RECIPIENTS => 10000,
 		self::CREATED_DAY => 0,
 		self::MAIL_LOG_DATA => 1000,
+		self::ID_ACTION_INDEX => 0,
 	];
 
 	public const array MIGRATION_SLEEP = [
 		self::MAIL_RECIPIENTS => 200000,
 		self::CREATED_DAY => 0,
 		self::MAIL_LOG_DATA => 200000,
+		self::ID_ACTION_INDEX => 200000,
 	];
 
 	public const array MIGRATION_HELP = [
 		self::MAIL_RECIPIENTS => "https://github.com/bilias/rqwatch/blob/master/docs/DB_MIGRATION.md",
 		self::CREATED_DAY => "https://github.com/bilias/rqwatch/blob/master/docs/DB_MIGRATION.md",
 		self::MAIL_LOG_DATA => "https://github.com/bilias/rqwatch/blob/master/docs/DB_MIGRATION.md",
+		self::ID_ACTION_INDEX => "https://github.com/bilias/rqwatch/blob/master/docs/DB_MIGRATION.md",
 	];
 
 	public const string STATUS_PENDING   = 'pending';