+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Master Dev Branch
 
+### 2026-08-06
+- Added id action Index Migration
+
 ### 2026-08-05
 - DB schema verifications related to migrations
 - Warm MigrationStatus cache in Kernel